pyproject.toml
[tool.poetry]
name = "classdoc"
version = "0.1.0"
description = "Generate docs from each file in your codebase."
authors = ["Ben Ranford <ben@ranford.dev>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
openai = "^1.52.2"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
